seperated rule translation
